refactored solo_xff_offset to leftmost_xff_address
