+++ docs/content/reference/api/github.com/solo-io/gloo/projects/gloo/api/external/envoy/extensions/xff_offset/solo_xff_offset_filter.proto.sk.md
@@ -25,7 +25,8 @@ weight: 5
 ---
 ### SoloXffOffset
 
-
+ 
+Envoy filter configuration for the solo_xff_offset filter,  not set by the user.
 
 ```yaml
 "offset": int
