+++ projects/gloo/api/external/envoy/extensions/xff_offset/solo_xff_offset_filter.proto
@@ -2,11 +2,12 @@ syntax = "proto3";
 
 package envoy.config.filter.http.solo_xff_offset.v2;
 
-option go_package = "github.com/solo-io/gloo/projects/gloo/pkg/api/external/envoy/extensions/solo_xff_offset";
+option go_package = "github.com/solo-io/gloo/projects/gloo/pkg/api/external/envoy/extensions/xff_offset";
 option java_package = "io.envoyproxy.envoy.config.filter.http.solo_xff_offset.v2";
 option java_outer_classname = "SoloXffOffsetProto";
 option java_multiple_files = true;
 
+// Envoy filter configuration for the solo_xff_offset filter,  not set by the user.
 message SoloXffOffset {
   // The amount of addresses from the left of x-forwarded-for (xff) header
   // which should be skipped to get the client address.
